+++ src/dbtype/write_document_response.rs
@@ -4,7 +4,7 @@ use DocumentId;
 use Revision;
 
 pub type DeleteDocumentResponse = WriteDocumentResponse;
-pub type PostToDatabaseResponse = WriteDocumentResponse;
+pub type PostDatabaseResponse = WriteDocumentResponse;
 pub type PutDocumentResponse = WriteDocumentResponse;
 
 #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]